seed: Depend on the seed before tarballing for development
Relying on an IFD-like behaviour is not possible with `fetchTarball`.

String inputs given to `fetchTarball` cannot depend on store paths via
string contexts. The alternative to rely on `unsafeDiscardStringContext`
works, but only if the store path already exists.

So demote into a more "end-to-end" like scenario that a developer can
opt into. Which would check that going from the produced tarball,
instead of effectively depending on a whole Nixpkgs, works.